requirements.txt
airtest==1.2.5
allure-pytest==2.9.45
pytest-xdist==2.5.0
pytest==7.1.2
click=8.1.3
pocoui==1.0.87
pytest-html==1.21.1
pytest-metadata==1.8.0
pytest-rerunfailures==7.0
pytest-parallel==0.1.1
suds-jurko==0.6